Compiler now passes per-suspend-point handle types
The <experimental/coroutine> header now defines some
concrete suspend-point-handle types that are parameterised
on a tag type that the compiler injects an anonymous type
for each suspend-point so that handles to different suspend
points are not convertible to each other.
Ideally the compiler would be auto-generating these types
to avoid type-erasure altogether, but this is as close as we
can get with the current LLVM intrinsics.

Then defined some general type-erasure mechanisms for
converting these per-suspend-point handle types to the
public type-erased handle types:
suspend_point_handle and continuation_handle.

The suspend_point_handle type now contains a pointer to the
frame and also a vtable pointer to the continuation state for
the current suspend-point. The vtable is constructed based
on the set of operations requested and passed as template
parameters to suspend_point_handle<>.