Eliminate duplicate calls to await_suspend() in co_await/co_yield expressions.
Now inserts an OpaqueValueExpr within expressions that invoke the continuation
returned from await_suspend() and ensures that this placeholder is replaced
with an lvalue reference to the return-value rather than repeating the
await_supsend() CallExpr in both the call to handle.__address() and the
call to handle.__continuation().

This required adding some extra sub-expressions to the CoroutineSuspendExpr.

Also removed the IsImplicit flag from CoawaitExpr as there are no longer
any implicit co_await expressions.

Note that similar changes still need to be applied to final-suspend expression
to avoid duplicate calls to promise.final_suspend() here too.